requirements.txt
# requirements.txt
fastapi
uvicorn[standard] # Includes standard dependencies like websockets, httptools
mysql-connector-python
groq
python-dotenv
pydantic
passlib[bcrypt] # For password hashing
python-jose[cryptography] # For JWT handling
email-validator # Often needed by Pydantic for EmailStr validation (good practice)